Create yabf version of `configfile init` command
