+++ src/main.js
@@ -1,10 +1,9 @@
 export * from './core/term-application'
+export * from './core/command'
 
-export * from './commands/scripts-list.command'
-export * from './commands/scripts-run.command'
-export * from './commands/modules-list.command'
-export * from './commands/modules-deploy.command'
+export * from './commands'
 
 export * from './services/file.service'
 export * from './services/exec/exec.service'
 export * from './services/deploy/deploy.service'
+export * from './services/config/config.service'
